backend/requirements.txt
#

# These requirements were autogenerated by pipenv

# To regenerate from the project's Pipfile, run:

#

#    pipenv lock --requirements

#



-i https://pypi.org/simple

asgiref==3.5.0; python_version >= '3.7'

beautifulsoup4==4.11.1

certifi==2021.10.8

cffi==1.15.0

charset-normalizer==2.0.12; python_version >= '3'

click==8.1.2; python_version >= '3.7'

colorama==0.4.4; platform_system == 'Windows'

cryptography==36.0.2; python_version >= '3.6'

deprecated==1.2.13; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'

django-cors-headers==3.11.0

django-oauth-toolkit==2.0.0

django==4.0.4

djangorestframework==3.13.1

gunicorn==20.1.0

h11==0.13.0; python_version >= '3.6'

idna==3.3; python_version >= '3'

jwcrypto==1.0

oauthlib==3.2.0; python_version >= '3.6'

pillow==9.1.0

psycopg2==2.9.3

pycparser==2.21

pytz==2022.1

requests==2.27.1

setuptools==62.1.0; python_version >= '3.7'

soupsieve==2.3.2.post1; python_version >= '3.6'

sqlparse==0.4.2; python_version >= '3.5'

tzdata==2022.1; sys_platform == 'win32'

urllib3==1.26.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'

uvicorn==0.17.6

wrapt==1.14.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'

